Moving ESP32Alarm from esp32serial to its own class
